# packages/api/openapi.yaml
openapi: 3.0.0
info:
  title: 1BRC for traP API
  version: 0.0.0
...
paths:
...
  /api/v1/submissions:
    post:
      operationId: createSubmission
      parameters: []
      responses:
        '202':
          description: The request has been accepted for processing, but processing has not yet completed.
          headers:
            location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionAccepted'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SubmissionUpload'
            encoding:
              executionKind:
                contentType: application/json
              language:
                contentType: application/json
              source:
                contentType: '*/*'
              binary:
                contentType: '*/*'
    get:
      operationId: listSubmissions
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionsResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
...
components:
  schemas:
...
    BenchmarkOutcome:
      type: object
      required:
        - verdict
        - scoreNs
        - error
      properties:
        verdict:
          $ref: '#/components/schemas/Verdict'
        scoreNs:
          type: string
          allOf:
            - $ref: '#/components/schemas/Nanoseconds'
          nullable: true
        error:
          type: string
          nullable: true
...
    DateTimeString:
      type: string
    ErrorDetail:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
    ExecutionKind:
      type: string
      enum:
        - native
        - javascript
        - typescript
        - bun
        - ruby
    Language:
      type: string
      enum:
        - c
        - cpp
        - go
        - rust
        - zig
        - csharp
        - other
        - javascript
        - typescript
        - bun
        - ruby
...
    Nanoseconds:
      type: string
      pattern: ^\d+$
...
    Submission:
      type: object
      required:
        - id
        - username
        - executionKind
        - language
        - sourceFilename
        - artifactSha256
        - status
        - public
        - infrastructureError
        - disqualifiedReason
        - uploadStartedAt
        - queuedAt
        - startedAt
        - completedAt
        - queueAhead
        - submissionNumber
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        username:
          type: string
        executionKind:
          allOf:
            - $ref: '#/components/schemas/ExecutionKind'
          nullable: true
        language:
          allOf:
            - $ref: '#/components/schemas/Language'
          nullable: true
        sourceFilename:
          type: string
          nullable: true
        artifactSha256:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        public:
          type: object
          allOf:
            - $ref: '#/components/schemas/BenchmarkOutcome'
          nullable: true
        private:
          type: object
          allOf:
            - $ref: '#/components/schemas/BenchmarkOutcome'
          nullable: true
        infrastructureError:
          type: string
          nullable: true
        disqualifiedReason:
          type: string
          nullable: true
        uploadStartedAt:
          $ref: '#/components/schemas/DateTimeString'
        queuedAt:
          type: string
          allOf:
            - $ref: '#/components/schemas/DateTimeString'
          nullable: true
        startedAt:
          type: string
          allOf:
            - $ref: '#/components/schemas/DateTimeString'
          nullable: true
        completedAt:
          type: string
          allOf:
            - $ref: '#/components/schemas/DateTimeString'
          nullable: true
        queueAhead:
          type: integer
          format: int32
          nullable: true
        submissionNumber:
          type: integer
          format: int32
          nullable: true
    SubmissionAccepted:
      type: object
      required:
        - id
        - status
        - statusUrl
        - uploadStartedAt
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum:
            - queued
        statusUrl:
          type: string
        uploadStartedAt:
          $ref: '#/components/schemas/DateTimeString'
...
    SubmissionStatus:
      type: string
      enum:
        - uploading
        - queued
        - running
        - completed
        - rejected
        - infrastructure_error
        - disqualified
    SubmissionUpload:
      type: object
      properties:
        executionKind:
          $ref: '#/components/schemas/ExecutionKind'
        language:
          $ref: '#/components/schemas/Language'
        source:
          type: string
          format: binary
        binary:
          type: string
          format: binary
      required:
        - executionKind
        - source
    SubmissionsResponse:
      type: object
      required:
        - submissions
      properties:
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
    Uuid:
      type: string
    Verdict:
      type: string
      enum:
        - accepted
        - wrong_answer
        - runtime_error
        - time_limit
        - output_limit
        - invalid_submission
        - infrastructure_error
        - disqualified